Refactor to a shared hook
